Revert "chore: reformat and clean CryptocurrenciesController and routes (#46)" (#47)
This reverts commit 4581b9a1a5be814aa2f1a083312624a0e2cfffca.

+++ api/app/Controllers/Http/CryptocurrenciesController.ts
@@ -4,24 +4,25 @@ import Cryptocurrency from '../../Models/Cryptocurrency'
 import { schema, validator } from '@ioc:Adonis/Core/Validator'
 export default class CryptocurrenciesController {
 
-    //List of Cryptocurrencies
-    public async index() {
+    //crud de la base de donnée des cryptocurrencies
+    // list of Cryptocurrencies
+    public async listCryptocurrencies() {
         return await Cryptocurrency.all()
     }
-
-    //Return one Cryptocurrency
-    public async show({ response, request }: HttpContextContract) {
+    // return one Cryptocurrency
+    public async oneCryptocurrencyBySlug({ params,response }: HttpContextContract) {
+        // const cryptocurrency = await Cryptocurrency.findBy('slug', params.slug)
         try {
-            return await Cryptocurrency.findByOrFail('slug', request.param('slug'));
-        } catch (e) {
-            response.notFound({ error: 'This slug is not affected to a Cryptocurrency' })
+            return await Cryptocurrency.findByOrFail('slug', params.slug)
+          } catch (e){
+            response.notAcceptable({error: 'This slug is not affected to a Cryptocurrency'})
             return null
-        }
+          }
     }
-
-    //Add a new Cryptocurrency
-    public async create({ request, response }: HttpContextContract) {
-        try {
+    // Add a new Cryptocurrency
+    public async newCryptocurrency({ request, response }: HttpContextContract): Promise<void> {
+        const existingCryptoVersion = await Cryptocurrency.findBy('slug', request.input('slug'))
+        if (!existingCryptoVersion) {
             const registerValidation = schema.create({
                 slug: schema.string(),
                 symbol: schema.string(),
@@ -41,43 +42,49 @@ export default class CryptocurrenciesController {
                 crypto,
             })
         }
-        catch (e) {
-            return response.notAcceptable({
+        else {
+            return response.ok({
                 message: 'A Cryptocurrency already exist with this slug'
             })
+
         }
     }
-
     //Delete One Cryptocurrency
-    public async destroy({ request, response }: HttpContextContract) {
-        try {
-            const cryptoToDelete = await Cryptocurrency.findByOrFail('slug', request.param('slug'))
+
+    public async deleteCryptocurrency({ request }: HttpContextContract) {
+
+        const cryptoToDelete = await Cryptocurrency.findBy('slug', request.input('slug'))
+        if (cryptoToDelete) {
             await cryptoToDelete.delete()
-            return response.ok({
-                message: 'Cryptocurrency successfully deleted'
-            })
-        } catch (e) {
-            response.notAcceptable({ error: 'Cryptocurrency deletion failed' })
-            return null
+            return { message: 'Profile successfully deleted' }
+        }
+        else {
+            return { message: 'Profile deletion failed' }
         }
     }
 
     //Update one Cryptocurrency
-    public async update({ request, response }: HttpContextContract) {
-        try {
-            const cryptocurrency = await Cryptocurrency.findByOrFail('slug', request.param('slug'))
-            cryptocurrency.slug = request.input('slug') ?? cryptocurrency.slug
-            cryptocurrency.symbol = request.input('symbol') ?? cryptocurrency.symbol
-            cryptocurrency.name = request.input('name') ?? cryptocurrency.name
-            cryptocurrency.sequence = request.input('sequence') ?? cryptocurrency.sequence
-            cryptocurrency.logo = request.input('logo') ?? cryptocurrency.logo
-            cryptocurrency.updatedAt = Date.now()
-            await cryptocurrency.save()
-            return { message: 'Profile updated successfully', cryptocurrency }
+    public async updateCryptocurrency({ request }: HttpContextContract) {
+        const cryptocurrency = await Cryptocurrency.findBy('slug', request.input('slug'))
+        if (cryptocurrency) {
 
-        } catch (e) {
-            response.notModified({ error: 'Cryptocurrency deletion failed' })
-            return null
+            try {
+                cryptocurrency.slug = request.input('newSlug') ?? cryptocurrency.slug
+                cryptocurrency.symbol = request.input('symbol') ?? cryptocurrency.symbol
+                cryptocurrency.name = request.input('name') ?? cryptocurrency.name
+                cryptocurrency.sequence = request.input('sequence') ?? cryptocurrency.sequence
+                cryptocurrency.logo = request.input('logo') ?? cryptocurrency.logo
+                cryptocurrency.updatedAt = Date.now()
+
+                await cryptocurrency?.save()
+                return { message: 'Profile updated successfully', cryptocurrency }
+            } catch (error) {
+                console.error('Profile update error:', error.message)
+                return { message: 'Profile update failed' }
+            }
+        }
+        else {
+            return { message: 'No profile find with this slug' }
         }
 
     }
